refactor: split ReviewPrompt into one component per file
